requirements.txt
boto3==1.35.90
botocore==1.35.90
jmespath==1.0.1
mysql-connector-python==9.1.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
